working on focusfiltereditor
trying to get focusFilterEditor to work along with storeFocusedElement to work properly. Currently the filter editor does not automatically get selected when the package is selected from the command pallate or the toggle. Esc or clicking out of the list view does not close the package either.

Added description to the listed themes